De-duplicate 15 jscpd clones across lint/refactor/semantic workspaces
