+++ src/lint/src/rules/gml/rules/prefer-hoistable-loop-accessors-rule.ts
@@ -1,28 +1,20 @@
+import * as CoreWorkspace from "@gmloop/core";
 import type { Rule } from "eslint";
 
 import type { GmlRuleDefinition } from "../../catalog.js";
 import {
     type AstNodeWithType,
     createMeta,
-    getNodeEndIndex,
     getNodeStartIndex,
-    isAstNodeRecord,
     isAstNodeWithType,
     walkAstNodes,
     walkAstNodesWithParent
 } from "../rule-base-helpers.js";
-import { isIdentifier, readObjectOption, shouldReportUnsafe } from "../rule-helpers.js";
+import { readObjectOption, shouldReportUnsafe } from "../rule-helpers.js";
 
 const DEFAULT_HOIST_ACCESSORS = Object.freeze({
     array_length: "len"
 });
-
-type LoopLengthAccessorCall = Readonly<{
-    functionName: string;
-    callStart: number;
-    callEnd: number;
-    callText: string;
-}>;
 
 type ForStatementContainerContext = Readonly<{
     forNode: AstNodeWithType;
@@ -67,90 +59,16 @@ function resolveSafeLocFromIndex(
     };
 }
 
-function resolveLoopLengthHoistSuffixMap(
-    functionSuffixOverrides: Record<string, string | null> | undefined
-): ReadonlyMap<string, string> {
-    const suffixMap = new Map<string, string>(Object.entries(DEFAULT_HOIST_ACCESSORS));
-    if (!functionSuffixOverrides) {
-        return suffixMap;
-    }
-
-    for (const [functionName, suffix] of Object.entries(functionSuffixOverrides)) {
-        if (!isIdentifier(functionName)) {
-            continue;
-        }
-
-        if (suffix === null) {
-            suffixMap.delete(functionName);
-            continue;
-        }
-
-        if (typeof suffix !== "string" || suffix.length === 0) {
-            continue;
-        }
-
-        suffixMap.set(functionName, suffix);
-    }
-
-    return suffixMap;
-}
-
-function collectLoopLengthAccessorCalls(parameters: {
-    sourceText: string;
-    rootNode: unknown;
-    enabledFunctionNames: ReadonlySet<string>;
-}): ReadonlyArray<LoopLengthAccessorCall> {
-    const collectedCalls: Array<LoopLengthAccessorCall> = [];
-    walkAstNodes(parameters.rootNode, (node) => {
-        if (!isAstNodeRecord(node) || node.type !== "CallExpression") {
-            return;
-        }
-
-        const callTarget = isAstNodeRecord(node.object) ? node.object : null;
-        if (
-            !callTarget ||
-            callTarget.type !== "Identifier" ||
-            typeof callTarget.name !== "string" ||
-            !parameters.enabledFunctionNames.has(callTarget.name)
-        ) {
-            return;
-        }
-
-        const start = getNodeStartIndex(node);
-        const end = getNodeEndIndex(node);
-        if (typeof start !== "number" || typeof end !== "number") {
-            return;
-        }
-
-        collectedCalls.push(
-            Object.freeze({
-                functionName: callTarget.name,
-                callStart: start,
-                callEnd: end,
-                callText: parameters.sourceText.slice(start, end)
-            })
-        );
-    });
-
-    return collectedCalls;
-}
-
 function collectLoopLengthAccessorCallsFromTestExpression(parameters: {
+    sourceText: string;
     testNode: unknown;
     enabledFunctionNames: ReadonlySet<string>;
-}): ReadonlyArray<LoopLengthAccessorCall> {
-    return collectLoopLengthAccessorCalls({
-        sourceText: "",
+}) {
+    return CoreWorkspace.Core.collectLoopLengthAccessorCallsFromAstNode({
+        sourceText: parameters.sourceText,
         rootNode: parameters.testNode,
         enabledFunctionNames: parameters.enabledFunctionNames
-    }).map((call) =>
-        Object.freeze({
-            functionName: call.functionName,
-            callStart: call.callStart,
-            callEnd: call.callEnd,
-            callText: call.callText
-        })
-    );
+    });
 }
 
 function collectForStatementContainerContexts(programNode: unknown): ReadonlyArray<ForStatementContainerContext> {
@@ -184,7 +102,10 @@ export function createPreferHoistableLoopAccessorsRule(definition: GmlRuleDefini
             const minOccurrences = typeof options.minOccurrences === "number" ? options.minOccurrences : 2;
             const functionSuffixes = options.functionSuffixes as Record<string, string | null> | undefined;
             const shouldReportUnsafeFixes = shouldReportUnsafe(context);
-            const suffixMap = resolveLoopLengthHoistSuffixMap(functionSuffixes);
+            const suffixMap = CoreWorkspace.Core.resolveIdentifierKeyedSuffixMap(
+                DEFAULT_HOIST_ACCESSORS,
+                functionSuffixes
+            );
 
             return Object.freeze({
                 Program(programNode) {
@@ -215,6 +136,7 @@ export function createPreferHoistableLoopAccessorsRule(definition: GmlRuleDefini
                     for (const loopNode of loopNodes) {
                         if (loopNode.type === "ForStatement" && enabledHoistFunctionNames.size > 0) {
                             const testCalls = collectLoopLengthAccessorCallsFromTestExpression({
+                                sourceText,
                                 testNode: (loopNode as any).test,
                                 enabledFunctionNames: enabledHoistFunctionNames
                             });
@@ -243,7 +165,7 @@ export function createPreferHoistableLoopAccessorsRule(definition: GmlRuleDefini
                             }
                         }
 
-                        const loopCalls = collectLoopLengthAccessorCalls({
+                        const loopCalls = CoreWorkspace.Core.collectLoopLengthAccessorCallsFromAstNode({
                             sourceText,
                             rootNode: loopNode,
                             enabledFunctionNames: new Set(["array_length"])
@@ -253,7 +175,7 @@ export function createPreferHoistableLoopAccessorsRule(definition: GmlRuleDefini
                         }
 
                         if (loopNode.type === "ForStatement") {
-                            const testCalls = collectLoopLengthAccessorCalls({
+                            const testCalls = CoreWorkspace.Core.collectLoopLengthAccessorCallsFromAstNode({
                                 sourceText,
                                 rootNode: (loopNode as any).test,
                                 enabledFunctionNames: new Set(["array_length"])
