+++ src/lint/test/rules/performance-regression.test.ts
@@ -1,22 +1,14 @@
 import assert from "node:assert/strict";
-import { createHash } from "node:crypto";
 import { readFile } from "node:fs/promises";
-import { test } from "node:test";
 
-import * as LintWorkspace from "@gmloop/lint";
-import { ESLint, type Linter } from "eslint";
-
-const { Lint } = LintWorkspace;
-const STILE_FIXTURE_URL = new URL("../../../../parser/test/input/stile.gml", import.meta.url);
-
-type TimedLintRunResult = Readonly<{
-    elapsedMilliseconds: number;
-    ruleMilliseconds: number;
-    messages: ReadonlyArray<ESLint.LintResult["messages"][number]>;
-    outputText: string;
-}>;
-
-const STILE_OPTIMIZE_MATH_OUTPUT_HASH = "42803788c231317796505783e423d1a02cdae11ac31648925faa6c3c51fa24f7";
+import {
+    buildLoopInvariantStressBatchSource,
+    createOutputHash,
+    lintSingleRuleWithTiming,
+    runSequentialPerformanceTest,
+    STILE_FIXTURE_URL,
+    STILE_OPTIMIZE_MATH_OUTPUT_HASH
+} from "./performance-test-helpers.js";
 
 function buildNonMathAssignmentBatchSource(statementCount: number): string {
     const lines: string[] = [];
@@ -74,26 +66,6 @@ function buildAdditiveIdentifierBatchSource(statementCount: number): string {
     return lines.join("\n");
 }
 
-function buildLoopInvariantStressBatchSource(loopCount: number, invariantTermsPerLoop: number): string {
-    const lines: string[] = [];
-
-    for (let loopIndex = 0; loopIndex < loopCount; loopIndex += 1) {
-        let invariantExpression = `(a_${loopIndex}_0 + b_${loopIndex}_0)`;
-        for (let termIndex = 1; termIndex < invariantTermsPerLoop; termIndex += 1) {
-            invariantExpression = `(${invariantExpression} + (a_${loopIndex}_${termIndex} + b_${loopIndex}_${termIndex}))`;
-        }
-
-        lines.push(
-            `repeat (count_${loopIndex}) {`,
-            `    total_${loopIndex} += (${invariantExpression}) + random(3);`,
-            "}"
-        );
-    }
-
-    lines.push("");
-    return lines.join("\n");
-}
-
 function buildLoopHoistCollisionStressSource(loopCount: number, reservedHoistNameCount: number): string {
     const lines: string[] = ["var cached_value = 0;"];
 
@@ -115,61 +87,15 @@ function buildLoopHoistCollisionStressSource(loopCount: number, reservedHoistNam
     return lines.join("\n");
 }
 
-async function lintSingleRuleWithTiming(
-    ruleId: string,
-    sourceText: string,
-    filePath = "performance-regression.gml"
-): Promise<TimedLintRunResult> {
-    const configEntry = {
-        files: ["**/*.gml"],
-        plugins: {
-            gml: Lint.plugin
-        },
-        language: "gml/gml",
-        rules: {
-            [ruleId]: "warn"
-        }
-    } satisfies Linter.Config;
-
-    const eslint = new ESLint({
-        overrideConfigFile: true,
-        fix: true,
-        stats: true,
-        overrideConfig: [configEntry]
-    });
-
-    const startedAtNanoseconds = process.hrtime.bigint();
-    const [result] = await eslint.lintText(sourceText, {
-        filePath
-    });
-    const elapsedMilliseconds = Number(process.hrtime.bigint() - startedAtNanoseconds) / 1e6;
-
-    const passTimings = result.stats?.times?.passes ?? [];
-    const ruleMilliseconds = passTimings.reduce((accumulator, passTiming) => {
-        return accumulator + (passTiming.rules[ruleId]?.total ?? 0);
-    }, 0);
-
-    return Object.freeze({
-        elapsedMilliseconds,
-        ruleMilliseconds,
-        messages: Object.freeze(result.messages),
-        outputText: result.output ?? sourceText
-    });
-}
-
-function createOutputHash(outputText: string): string {
-    return createHash("sha256").update(outputText).digest("hex");
-}
-
-function runSequentialPerformanceTest(name: string, implementation: () => Promise<void>): void {
-    void test(name, { concurrency: false }, implementation);
-}
-
 runSequentialPerformanceTest(
     "optimize-math-expressions skips non-math batches without runaway traversal cost",
     async () => {
         const source = buildNonMathAssignmentBatchSource(1500);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-math-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-math-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.equal(timedRun.outputText, source);
@@ -188,7 +114,11 @@ runSequentialPerformanceTest(
     "optimize-logical-flow skips non-logical batches without deep clone overhead",
     async () => {
         const source = buildNonLogicalConditionBatchSource(1200);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-logical-flow", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-logical-flow",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.equal(timedRun.outputText, source);
@@ -207,7 +137,11 @@ runSequentialPerformanceTest(
     "optimize-logical-flow avoids deep-cloning large guard bodies that cannot be simplified",
     async () => {
         const source = buildHeavyIfGuardBatchSource(300);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-logical-flow", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-logical-flow",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.equal(timedRun.outputText, source);
@@ -226,7 +160,11 @@ runSequentialPerformanceTest(
     "optimize-math-expressions scales linearly for long arithmetic assignment batches",
     async () => {
         const source = buildArithmeticChainBatchSource(250);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-math-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-math-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.ok(
@@ -248,7 +186,11 @@ runSequentialPerformanceTest(
     "optimize-math-expressions keeps dot-product auto-fixes within bounded runtime on large batches",
     async () => {
         const source = buildArithmeticChainBatchSource(1000);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-math-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-math-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.ok(
@@ -270,7 +212,11 @@ runSequentialPerformanceTest(
     "optimize-math-expressions skips additive identifier batches without clone-heavy normalization",
     async () => {
         const source = buildAdditiveIdentifierBatchSource(2500);
-        const timedRun = await lintSingleRuleWithTiming("gml/optimize-math-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/optimize-math-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.equal(timedRun.outputText, source);
@@ -304,7 +250,11 @@ runSequentialPerformanceTest(
     "prefer-loop-invariant-expressions avoids repeated subtree analysis on deep invariant loop expressions",
     async () => {
         const source = buildLoopInvariantStressBatchSource(60, 15);
-        const timedRun = await lintSingleRuleWithTiming("gml/prefer-loop-invariant-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/prefer-loop-invariant-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.ok(
@@ -326,7 +276,11 @@ runSequentialPerformanceTest(
     "prefer-loop-invariant-expressions keeps large hoist-name resolution workloads within bounded runtime",
     async () => {
         const source = buildLoopInvariantStressBatchSource(160, 30);
-        const timedRun = await lintSingleRuleWithTiming("gml/prefer-loop-invariant-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/prefer-loop-invariant-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.ok(
@@ -374,7 +328,11 @@ runSequentialPerformanceTest(
     "prefer-loop-invariant-expressions keeps very large hoist-name resolution workloads within bounded runtime",
     async () => {
         const source = buildLoopInvariantStressBatchSource(320, 60);
-        const timedRun = await lintSingleRuleWithTiming("gml/prefer-loop-invariant-expressions", source);
+        const timedRun = await lintSingleRuleWithTiming(
+            "gml/prefer-loop-invariant-expressions",
+            source,
+            "performance-regression.gml"
+        );
 
         assert.equal(timedRun.messages.length, 0);
         assert.ok(
