+++ src/refactor/test/fixtures/loop-length-hoisting-basic/expected/main.gml
@@ -1,0 +1,4 @@
+var len = array_length(items);
+for (var i = 0; i < len; i++) {
+    total += i;
+}
