+++ src/refactor/test/fixtures/loop-length-hoisting-basic/project/main.gml
@@ -1,0 +1,3 @@
+for (var i = 0; i < array_length(items); i++) {
+    total += i;
+}
