+++ src/refactor/test/fixtures/naming-convention-cross-file/expected/definitions.gml
@@ -1,0 +1,3 @@
+function myFunction() {
+    show_debug_message("hello");
+}
