+++ src/refactor/test/fixtures/naming-convention-cross-file/expected/usage.gml
@@ -1,0 +1,1 @@
+myFunction();
