+++ src/refactor/test/fixtures/naming-convention-cross-file/project/definitions.gml
@@ -1,0 +1,3 @@
+function my_function() {
+    show_debug_message("hello");
+}
