+++ src/refactor/test/fixtures/naming-convention-cross-file/project/usage.gml
@@ -1,0 +1,1 @@
+my_function();
